refactor: util module
