+++ llrt_core/src/modules/util/text_encoder.rs
